+++ services/libs/snowflake/src/snowflakeExporter.ts
@@ -25,6 +25,22 @@ interface CopyIntoRow {
   fileName: string
 }
 
+/**
+ * Build the S3 filename prefix for a batched COPY INTO export.
+ * Format: {CROWD_SNOWFLAKE_S3_BUCKET_PATH}/{platform}/{sourceName}/{yyyy}/{mm}/{dd}
+ */
+export function buildS3FilenamePrefix(platform: string, sourceName: string): string {
+  const s3BucketPath = process.env.CROWD_SNOWFLAKE_S3_BUCKET_PATH
+  if (!s3BucketPath) {
+    throw new Error('Missing required env var CROWD_SNOWFLAKE_S3_BUCKET_PATH')
+  }
+  const now = new Date()
+  const year = now.getFullYear()
+  const month = String(now.getMonth() + 1).padStart(2, '0')
+  const day = String(now.getDate()).padStart(2, '0')
+  return `${s3BucketPath}/${platform}/${sourceName}/${year}/${month}/${day}`
+}
+
 export class SnowflakeExporter {
   private readonly snowflake: SnowflakeClient
